Se corrige forma de llamado, se agrega seguridad, SE AGREGA PETICION POR ID

+++ Backend/FondosVoluntarios/src/main/java/com/co/fondovoluntario/FondosVoluntarios/dominio/nuevosfondos/puerto/persistencia/RepositorioCrearFondo.java
@@ -9,7 +9,7 @@ public interface RepositorioCrearFondo {
 
     Optional<Boolean> guardar(Fondos fondo);
 
-    Optional<Fondos> consultarXId(Long id);
+    Optional<Fondos> consultarXId(Integer id);
 
     Optional<List<Fondos>> consultarTodos();
 }
