+++ Backend/FondosVoluntarios/src/main/java/com/co/fondovoluntario/FondosVoluntarios/infraestructura/nuevosfondos/servicios/ServicioFondosFPV.java
@@ -22,7 +22,7 @@ public class ServicioFondosFPV {
         return manejadorFondos.consultar().orElse(null);
     }
 
-    public DtoFondo consultarXId(Long id) {
+    public DtoFondo consultarXId(Integer id) {
         return manejadorFondos.consultarXId(id).orElse(null);
     }
 }
